+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 - Increase number of imported radars by fetch the next batch #6 Idea thanks to @futuretap
+- Improve paste email to the import textField #10
 
 ## [1.0.0(7)] - 2018-07-09
 ### Added